added custom hook for fetching crewid, for reuse in footer links, updated styling and deleted phone and twitter in footer com
