+++ src/components/footer/SosialMediaLinks.tsx
@@ -7,16 +7,13 @@ const SocialMediaLinks = () => {
     <div className={styles["container-sosial-media"]}>
       <HeadingThree>Follow us?</HeadingThree>
       <div className={styles["wrapper-sosial-icon"]}>
-        <a href="https://www.instagram.com/">
+        <a href="https://www.instagram.com/original_vikings_of_norway">
           <FaInstagram className={styles["social-icon"]} />
         </a>
-        <a href="https://twitter.com/">
-          <FaTwitter className={styles["social-icon"]} />
-        </a>
-        <a href="https://www.youtube.com/">
+        <a href="https://www.youtube.com/channel/UCaPUAvRBw0i5ET79TMh2_MQ">
           <FaYoutube className={styles["social-icon"]} />
         </a>
-        <a href="https://www.facebook.com/">
+        <a href="https://www.facebook.com/VikingskipetSagaFarmann">
           <FaFacebook className={styles["social-icon"]} />
         </a>
       </div>
