+++ src/components/navigation/hamburger/HamburgerTransition.tsx
@@ -1,37 +1,16 @@
-import React, { useState, useRef, useEffect } from "react";
+import React, { useState, useRef } from "react";
 import { CSSTransition } from "react-transition-group";
 import styles from "./HamburgerTransition.module.css";
 import Link from "next/link";
 import { IoIosMenu, IoMdClose } from "react-icons/io";
-import API_ENDPOINTS from "@/endpoints/endpoints";
-import { CrewMember } from "@/pages/crew_members/[id]";
+import { useCurrentCrewId } from "@/hooks/useCurrentCrewId";
 
 const HamburgerTransition = () => {
   const [isOpen, setIsOpen] = useState(false);
   const [click, setClick] = useState(false);
   const nodeRefBurger = useRef<HTMLButtonElement>(null);
   const nodeRefMenu = useRef<HTMLDivElement>(null);
-  const [crewId, setCrewId] = useState();
-
-  useEffect(() => {
-    const getCurrentCrew = async () => {
-      const [resCrewMembers] = await Promise.all([
-        fetch(API_ENDPOINTS.crewMembers),
-      ]);
-      const [crewIds] = await Promise.all([resCrewMembers.json()]);
-
-      const currentCrewId = crewIds.filter(
-        (crew: CrewMember) => crew.acf?.current_crew === true
-      );
-
-      const currentCrew = currentCrewId.map((crew: CrewMember) =>
-        crew.id.toString()
-      );
-      console.log(currentCrew);
-      setCrewId(currentCrew);
-    };
-    getCurrentCrew();
-  }, []);
+  const crewId = useCurrentCrewId();
 
   const toggleMenu = () => {
     setIsOpen((prevState) => !prevState);
